fix: use DOUBLE PRECISION instead of DOUBLE in V62 migration
PostgreSQL does not support bare DOUBLE type — requires DOUBLE PRECISION.

+++ backend/src/main/resources/db/migration/V62__add_sky_brightness_sqm_to_locations.sql
@@ -1,4 +1,4 @@
 -- Add SQM (Sky Quality Meter) value alongside the existing Bortle class.
 -- SQM is a continuous value (magnitudes per square arcsecond) providing finer
 -- granularity than the discrete Bortle 1–8 scale.
-ALTER TABLE locations ADD COLUMN IF NOT EXISTS sky_brightness_sqm DOUBLE;
+ALTER TABLE locations ADD COLUMN IF NOT EXISTS sky_brightness_sqm DOUBLE PRECISION;
